Revert "Migrated to WPF, refactored monolith class to several, light MVVM lib"
This reverts commit 369e36ecc965b33cfb323082993368b73f69f899.
