+++ JitMagic/JitMagic.sln
@@ -1,25 +1,25 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.11.35125.118
+# Visual Studio 15
+VisualStudioVersion = 15.0.27520.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JitMagic", "JitMagic\JitMagic.csproj", "{5FB5914D-E98B-46A5-A8FE-2B3294CE65EF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JitMagic", "JitMagic\JitMagic.csproj", "{63A0047E-64E8-4620-BB50-AE1CA17B3975}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Release|x64 = Release|x64
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5FB5914D-E98B-46A5-A8FE-2B3294CE65EF}.Debug|x64.ActiveCfg = Debug|x64
-		{5FB5914D-E98B-46A5-A8FE-2B3294CE65EF}.Debug|x64.Build.0 = Debug|x64
-		{5FB5914D-E98B-46A5-A8FE-2B3294CE65EF}.Release|x64.ActiveCfg = Release|x64
-		{5FB5914D-E98B-46A5-A8FE-2B3294CE65EF}.Release|x64.Build.0 = Release|x64
+		{63A0047E-64E8-4620-BB50-AE1CA17B3975}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63A0047E-64E8-4620-BB50-AE1CA17B3975}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63A0047E-64E8-4620-BB50-AE1CA17B3975}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63A0047E-64E8-4620-BB50-AE1CA17B3975}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B764A2FB-48BD-4791-8E8A-F69DC5D718AE}
+		SolutionGuid = {D959A076-86C3-4F29-BF66-6ECD4D611E2D}
 	EndGlobalSection
 EndGlobal
