--- JitMagic/JitMagic/App.xaml
@@ -1,9 +1,0 @@
-﻿<Application x:Class="JitMagic.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:JitMagic"
-             StartupUri="Views/JITSelectorWindow.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
