--- JitMagic/JitMagic/App.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Configuration;
-using System.Data;
-using System.Windows;
-
-namespace JitMagic {
-	/// <summary>
-	/// Interaction logic for App.xaml
-	/// </summary>
-	public partial class App : Application {
-	}
-
-}
