+++ JitMagic/JitMagic/NativeMethods.txt
