--- JitMagic/JitMagic/Views/JITSelectorWindow.xaml
@@ -1,82 +1,0 @@
-<Window x:Class="JitMagic.Views.JITSelectorWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:JitMagic" xmlns:vm="clr-namespace:JitMagic.ViewModels" Visibility="Collapsed" d:Visibility="Visible"
-		mc:Ignorable="d" Topmost="{Binding TopMost,Mode=OneWay}" Background="#FFECEBEB" WindowStartupLocation="CenterScreen"
-        Title="{Binding WindowTitle}" Height="{Binding WinHeight,Mode=TwoWay}" Width="{Binding WinWidth,Mode=TwoWay}" Icon="pack://application:,,,/JitMagic.ico">
-	<Window.DataContext>
-		<vm:JITSelectorViewModel/>
-	</Window.DataContext>
-	<Window.Resources>
-		<Style BasedOn="{StaticResource {x:Type Button}}" TargetType="{x:Type Button}">
-			<Setter Property="Margin" Value="5" />
-			<Setter Property="Padding" Value="3" />
-		</Style>
-	</Window.Resources>
-	<Grid Margin="5" >
-		<Grid.RowDefinitions>
-			<RowDefinition Height="Auto"/>
-			<RowDefinition  />
-			<RowDefinition Height="Auto"/>
-		</Grid.RowDefinitions>
-		<TextBlock Margin="5,5,5,0" Text="{Binding ProcessInfo}" />
-		<ListBox x:Name="listDebuggers" IsSynchronizedWithCurrentItem="True" KeyDown="ListBox_KeyDown" Margin="5" ItemsSource="{Binding debuggers}" SelectedItem="{Binding selected_debugger, Mode=TwoWay}" ScrollViewer.HorizontalScrollBarVisibility="Disabled"  ScrollViewer.VerticalScrollBarVisibility="Auto" Grid.Row="1" >
-			<ListBox.Resources>
-				<Style TargetType="{x:Type ListBoxItem}" BasedOn="{StaticResource {x:Type ListBoxItem}}" >
-					<Setter Property="Margin" Value="5,5,5,0" />
-				</Style>
-			</ListBox.Resources>
-			<ListBox.ItemsPanel>
-				<ItemsPanelTemplate>
-					<WrapPanel MinHeight="80" Orientation="Horizontal" IsItemsHost="True" />
-				</ItemsPanelTemplate>
-			</ListBox.ItemsPanel>
-			<ListBox.ItemTemplate>
-				<DataTemplate>
-					<Grid Width="75" Height="80">
-						<Grid.InputBindings>
-							<MouseBinding MouseAction="LeftDoubleClick" Command="{Binding DataContext.DebuggerDoubleClickedCmd, RelativeSource={RelativeSource AncestorType={x:Type ListBox}}}"/>
-						</Grid.InputBindings>
-						<Grid.RowDefinitions>
-							<RowDefinition Height="30"/>
-							<RowDefinition />
-						</Grid.RowDefinitions>
-						<Image Source="{Binding DisplayIcon}" MaxWidth="60" />
-						<TextBlock Grid.Row="1" Text="{Binding Name}" Margin="0,5,0,0" TextAlignment="Center" TextWrapping="WrapWithOverflow" FontSize="13"/>
-					</Grid>
-				</DataTemplate>
-			</ListBox.ItemTemplate>
-
-		</ListBox>
-		<!--</ScrollViewer>-->
-		<Grid Grid.Row="2">
-			<Grid.ColumnDefinitions>
-				<ColumnDefinition Width="{Binding LeftCommandColWidth}" />
-				<ColumnDefinition Width="Auto" />
-				<ColumnDefinition Width="*" />
-			</Grid.ColumnDefinitions>
-			<StackPanel Orientation="Horizontal">
-				<Button Content="Ignore All" Command="{Binding IgnoreAllCmd}" />
-				<TextBlock Text="for "  VerticalAlignment="Center" />
-				<TextBox VerticalAlignment="Center" Width="20" Text="{Binding IgnoreForMinutes, Mode=TwoWay}" />
-				<TextBlock  VerticalAlignment="Center" Text=" minutes" />
-
-			</StackPanel>
-			<StackPanel Orientation="Horizontal" Grid.Column="1" HorizontalAlignment="Center">
-				<Button Content="{Binding AttachText}" IsDefault="True" Command="{Binding AttachCmd}" />
-			</StackPanel>
-			<ScrollViewer  Grid.Column="2" VerticalScrollBarVisibility="Disabled" HorizontalScrollBarVisibility="Auto">
-				<StackPanel Orientation="Horizontal" HorizontalAlignment="Right">
-					<Button Content="Set Window Size As Default" Command="{Binding SaveWindowSizeCmd}" Visibility="{Binding StandardLaunchOnlyVisibility}" />
-					<Button Content="Remove This Debugger" Command="{Binding RemoveSelectedDebuggerCmd}" Visibility="{Binding StandardLaunchOnlyVisibility}" />
-					<Button Content="Blacklist This App" Command="{Binding BlacklistAppCmd}" Visibility="{Binding AEDebugOnlyVisibility}" />
-					<Button Content="Remove As JIT" Command="{Binding RemoveAsJITCmd}" Visibility="{Binding StandardLaunchOnlyVisibility}" />
-					<Button Content="Launch W/O JIT For Options" Command="{Binding LaunchNormalCmd}" Visibility="{Binding AEDebugOnlyVisibility}" />
-				</StackPanel>
-			</ScrollViewer>
-
-		</Grid>
-	</Grid>
-</Window>
